Project Completed

+++ lib/Models/DataModel.dart
@@ -9,7 +9,7 @@ class PatientsData {
   int? active;
   int? critical;
   int? casesPerOneMillion;
-  double? deathsPerOneMillion;
+  int? deathsPerOneMillion;
   int? tests;
   double? testsPerOneMillion;
   int? population;
